BREAKING CHANGE: add Docker support
This commit adds support for Docker to SoundWave. The project now includes a Dockerfile and a docker-compose.yml file for building and running the application in a Docker container.

The Docker image is based on the official Python 3.9 image and includes all the required dependencies to run the Django application. The Dockerfile also copies the application code into the image and exposes the necessary ports for running the application.

With this change, the application can be easily deployed and run in a Docker environment, making it more portable and easier to manage.